+++ source/arm9/keyboard.c
@@ -257,7 +257,7 @@ void keyboardUpdate(void)
    }
 }
 
-int keyboardRead(struct _reent *r, int unused, char *ptr, int len)
+ssize_t keyboardRead(struct _reent *r, int unused, char *ptr, size_t len)
 {	
    int wasHidden = 0;
    int tempLen;
